created space for 'coders,' -
Facebook’s algorithms shaping the news. Self-driving cars roaming the streets. Revolution on Twitter and romance on Tinder. We live in a world constructed of code–and coders are the ones who built it for us. From acclaimed tech writer Clive Thompson comes a brilliant anthropological reckoning with the most powerful tribe in the world today, computer programmers, in a book that interrogates who they are, how they think, what qualifies as greatness in their world, and what should give us pause. They are the most quietly influential people on the planet, and Coders shines a light on their culture.

In pop culture and media, the people who create the code that rules our world are regularly portrayed in hackneyed, simplified terms, as ciphers in hoodies. Thompson goes far deeper, dramatizing the psychology of the invisible architects of the culture, exploring their passions and their values, as well as their messy history. In nuanced portraits, Coders takes us close to some of the great programmers of our time, including the creators of Facebook’s News Feed, Instagram, Google’s cutting-edge AI, and more. Speaking to everyone from revered “10X” elites to neophytes, back-end engineers and front-end designers, Thompson explores the distinctive psychology of this vocation–which combines a love of logic, an obsession with efficiency, the joy of puzzle-solving, and a superhuman tolerance for mind-bending frustration.

Along the way, Coders thoughtfully ponders the morality and politics of code, including its implications for civic life and the economy. Programmers shape our everyday behavior: When they make something easy to do, we do more of it. When they make it hard or impossible, we do less of it. Thompson wrestles with the major controversies of our era, from the “disruption” fetish of Silicon Valley to the struggle for inclusion by marginalized groups.